internal/envoy: Refactor the regex file into the v2 package
The matcher package is for v2 xDS resource APIs, so needs to be in the internal/envoy/v2 package.
