+++ internal/envoy/v2/regex.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package envoy
+package v2
 
 import (
 	matcher "github.com/envoyproxy/go-control-plane/envoy/type/matcher"
